nir: fix input/output info gathering for lowered IO
Ooops.

Fixes: 17af07024df - nir: gather all IO info from IO intrinsics

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/6758>

+++ src/compiler/nir/nir_gather_info.c
@@ -308,7 +308,7 @@ static void
 gather_intrinsic_info(nir_intrinsic_instr *instr, nir_shader *shader,
                       void *dead_ctx)
 {
-   unsigned slot_mask = 0;
+   uint64_t slot_mask = 0;
 
    if (nir_intrinsic_infos[instr->intrinsic].index_map[NIR_INTRINSIC_IO_SEMANTICS] > 0) {
       nir_io_semantics semantics = nir_intrinsic_io_semantics(instr);
